Reverse Order
Write a function that takes a string as input, and returns a string similar to the input, but with the words in reverse
order, and the punctuation marks maintaining their original order.
E.g.:
f("Hello. I'm Edwin A.J, and you?") => "You. and A.J Edwin I'm, Hello?"
f("What time is it? Hammer time.") => "Time Hammer It is? time what."

Note: As shown in the example above, the order of the punctuation marks ("?", "," , ".") have not changed. Only the words have.