Man in the Well
A man is stuck at the bottom of a well. Each day, he climbs up 8 metres, and then at night, he slips 3 metres. Using loop (any loop of your choice), write a function to determine and print how long he takes to climb out of a well of any given depth, where the depth of the well is taken as input.