Caesar Cipher
A Caesar cipher is an ancient form of encryption. It involves taking a text (s string) as input, an encoding it by replacing each letter by one n-steps next to the alphabet.

Write a function that takes a plaintext (string) to encode, and a shift value and output the encoded value of the string

Write another function that takes in the encoded string, with a shift value and decodes it.

Finally, write a third function that takes in a text, a shift value, a third parameter to indicate whether to encode or decode the given text (i.e., f("string", 5, True/False)), and print out the encoded (or decoded) text accordingly.