Sudoku Solver
* Write a function that takes in a 9×9 array of numbers.
Let this list represent a partially filled grid of numbers (specifically, of integers ranging from 1 to 9,
where "0" signifies an empty space in the grid.) as parameters, and returns as a 9×9 array, the solution(s) to it. If there are no solutions, return False.

Exceptions:
* If an empty list or an invalid list (list with numbers outside the 1-9 range, or empty lists, or a list of wrong size, etc.) is input, it issues a warning.
(Hint: Backtracking is one effective method for solving this problem—it is however, not the only method.)

Given a partially filled 9×9 list, the goal is to assign digits (from 1 to 9) to the empty cells so that every row, column, and sub grid of size 3×3 contains exactly one instance of the digits from 1 to 9.